fix: BoxHashtag now it is a Component
