+++ src/pages/Timeline.js
@@ -8,13 +8,11 @@ import PageContainer from "./../components/PageContainer";
 import CreatePost from "./../components/CreatePost";
 import Post from "./../components/Post";
 import requestPostsApi from "./../services/api/posts";
-import requestHashtagsApi from "../services/api/hashtags";
 import styled from "styled-components";
 
 function Timeline() {
   const [loading, setLoading] = useState(true);
   const [data, setData] = useState([]);
-  const [hashtags, setHashtags] = useState({});
   const [error, setError] = useState(false);
   const [updatePage, setUpdatePage] = useState(0);
   const [newPosts, setNewPosts] = useState(false);
@@ -48,16 +46,6 @@ function Timeline() {
     promise.catch((error) => {
       setError(true);
       setLoading(false);
-      console.log(error.message);
-    });
-
-    const promiseHashtag = requestHashtagsApi.getHashtags();
-    promiseHashtag.then((response) => {
-      setHashtags(response.data);
-      //setLoading(false);
-    });
-    promiseHashtag.catch((error) => {
-      //setLoading(false);
       console.log(error.message);
     });
   }, [updatePage]);
@@ -145,26 +133,6 @@ function Timeline() {
               )}
             </InfiniteScroll>
           </div>
-          <ContainerHashtag>
-            <div>
-              <h1>trending</h1>
-            </div>
-            {loading ? (
-              <p>Loading...</p>
-            ) : hashtags.length > 0 ? (
-              hashtags.map((hashtag) => (
-                <p
-                  onClick={() => {
-                    navigate(`/hashtag/${hashtag.name}`);
-                  }}
-                >
-                  # {hashtag.name}
-                </p>
-              ))
-            ) : (
-              <p>There are no hashtags yet</p>
-            )}
-          </ContainerHashtag>
         </DivFlex>
       </PageContainer>
     </>
@@ -175,45 +143,6 @@ export default Timeline;
 
 const DivFlex = styled.div`
   display: flex;
-`;
-
-const ContainerHashtag = styled.div`
-  margin-left: 25px;
-  width: 301px;
-  height: 406px;
-  background: #171717;
-  border-radius: 16px;
-  color: #ffffff;
-  font-style: normal;
-  font-weight: 700;
-
-  div {
-    padding-bottom: 12px;
-    margin-bottom: 22px;
-    border-bottom: 1px solid #484848;
-  }
-
-  h1 {
-    font-family: "Oswald";
-    font-size: 27px;
-    line-height: 40px;
-    margin-left: 16px;
-    margin-top: 8px;
-  }
-
-  p {
-    font-family: "Lato";
-    font-style: normal;
-    font-weight: 700;
-    font-size: 19px;
-    line-height: 23px;
-    letter-spacing: 0.05em;
-    margin-left: 16px;
-  }
-
-  @media only screen and (max-width: 800px) {
-    display: none;
-  }
 `;
 
 const BlueBox = styled.div`
